+++ internal/c2go/h3index__cellToChildPos_parity_test.go
@@ -7,7 +7,7 @@ import "testing"
 func Test_h3index_cellToChildPos_ParityWithC(t *testing.T) {
 	hs := []H3Index{0x8928308280fffff, 0x821c07fffffffff}
 	for _, child := range hs {
-		for parentRes := 0; parentRes <= getResolution(child); parentRes++ {
+		for parentRes := int32(0); parentRes <= getResolution(child); parentRes++ {
 			goPos, goErr := cellToChildPos(child, parentRes)
 			cPos, cErr := cellToChildPosC(child, parentRes)
 			if goErr != H3Error(cErr) || goPos != cPos {
