+++ internal/c2go/h3index__childPosToCell.go
@@ -3,7 +3,7 @@ package c2go
 // childPosToCell returns the child cell at a given position under parent at childRes.
 // Ports H3_EXPORT(childPosToCell).
 // Ported from H3 C: h3Index.c::childPosToCell
-func childPosToCell(childPos int64, parent H3Index, childRes int) (H3Index, H3Error) {
+func childPosToCell(childPos int64, parent H3Index, childRes int32) (H3Index, H3Error) {
 	if childRes < 0 || childRes > MAX_H3_RES {
 		return 0, E_RES_DOMAIN
 	}
@@ -27,7 +27,7 @@ func childPosToCell(childPos int64, parent H3Index, childRes int) (H3Index, H3Er
 
 	if isPentagon(parent) {
 		inPent := true
-		for res := 1; res <= resOffset; res++ {
+		for res := int32(1); res <= resOffset; res++ {
 			resWidth := _ipow(7, int64(resOffset-res))
 			if inPent {
 				pentWidth := 1 + (5*(resWidth-1))/6
@@ -36,18 +36,18 @@ func childPosToCell(childPos int64, parent H3Index, childRes int) (H3Index, H3Er
 				} else {
 					idx -= pentWidth
 					inPent = false
-					child = setIndexDigit(child, parentRes+res, int((idx/resWidth)+2))
+					child = setIndexDigit(child, parentRes+res, int32((idx/resWidth)+2))
 					idx %= resWidth
 				}
 			} else {
-				child = setIndexDigit(child, parentRes+res, int(idx/resWidth))
+				child = setIndexDigit(child, parentRes+res, int32(idx/resWidth))
 				idx %= resWidth
 			}
 		}
 	} else {
-		for res := 1; res <= resOffset; res++ {
+		for res := int32(1); res <= resOffset; res++ {
 			resWidth := _ipow(7, int64(resOffset-res))
-			child = setIndexDigit(child, parentRes+res, int(idx/resWidth))
+			child = setIndexDigit(child, parentRes+res, int32(idx/resWidth))
 			idx %= resWidth
 		}
 	}
